Testes com findAll, save e findById

+++ backend/src/main/java/com/nexola/dscatalog/services/ProductService.java
@@ -62,6 +62,9 @@ public class ProductService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         try {
             repository.deleteById(id);
         }
